requirements.txt
cython
scipy
tqdm
mgmetis
matplotlib
pyacvd
h5py
pyvista
numpy
lxml
networkx
pandas
PyQt5
pytest
PyYAML
schema
vtk
plotly
